refactor: rename fielddef to channeldef
