+++ packages/superset-ui-preset-chart-xy/src/encodeable/AxisAgent.ts
@@ -5,7 +5,7 @@ import { getTextDimension } from '@superset-ui/dimension';
 import { CategoricalColorScale } from '@superset-ui/color';
 import { extractFormatFromTypeAndFormat } from './parsers/extractFormat';
 import { CoreAxis, LabelOverlapStrategy } from './types/Axis';
-import { PositionFieldDef, ChannelDef } from './types/FieldDef';
+import { PositionFieldDef, ChannelDef } from './types/ChannelDef';
 import ChannelEncoder from './ChannelEncoder';
 import { DEFAULT_LABEL_ANGLE } from '../utils/constants';
