+++ packages/superset-ui-preset-chart-xy/src/encodeable/types/Channel.ts
@@ -1,5 +1,5 @@
 import { Value } from 'vega-lite/build/src/channeldef';
-import { XFieldDef, YFieldDef, ChannelDef, MarkPropChannelDef, TextChannelDef } from './FieldDef';
+import { XFieldDef, YFieldDef, ChannelDef, MarkPropChannelDef, TextChannelDef } from './ChannelDef';
 import { ObjectWithKeysFromAndValueType } from './Base';
 
 // eslint-disable-next-line import/prefer-default-export
